+++ include/Small/Widgets.hpp
@@ -2,7 +2,7 @@
  * @file Widgets.hpp
  * @author Alexandre Plateau (lexplt.dev@gmail.com)
  * @brief Include all the files from the Widgets folder
- * @version 0.2
+ * @version 0.3
  * @date 2020-04-15
  * 
  * @copyright Copyright (c) 2020
@@ -20,5 +20,9 @@ namespace sgl::Widgets {}
 
 #include <Small/Widgets/Base.hpp>
 #include <Small/Widgets/Layout.hpp>
+#include <Small/Widgets/Button.hpp>
+#include <Small/Widgets/CheckBox.hpp>
+#include <Small/Widgets/TextBox.hpp>
+#include <Small/Widgets/Label.hpp>
 
 #endif