+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added `Small/System.hpp` and `Small/Scenes.hpp` to `Small/All.hpp`
 - adding a base widget to be completed
 - adding a layout widget to hold many widgets
+- adding a button widget
 
 ### Changed
 - the tilemap has been updated to no longer hold the texture, and be able to modify the vertex array when needed